style(security): move security package to root
