+++ src/main/java/com/e2i/wemeet/dto/response/credential/SmsCredentialResponse.java
@@ -1,7 +1,7 @@
 package com.e2i.wemeet.dto.response.credential;
 
-import com.e2i.wemeet.config.security.model.MemberPrincipal;
 import com.e2i.wemeet.domain.member.RegistrationType;
+import com.e2i.wemeet.security.model.MemberPrincipal;
 import java.util.Collection;
 import org.springframework.security.core.GrantedAuthority;
