+++ src/main/java/com/e2i/wemeet/security/handler/CustomAuthenticationSuccessHandler.java
@@ -1,9 +1,9 @@
-package com.e2i.wemeet.config.security.handler;
+package com.e2i.wemeet.security.handler;
 
-import com.e2i.wemeet.config.security.model.MemberPrincipal;
-import com.e2i.wemeet.config.security.token.TokenInjector;
 import com.e2i.wemeet.dto.response.ResponseDto;
 import com.e2i.wemeet.dto.response.credential.SmsCredentialResponse;
+import com.e2i.wemeet.security.model.MemberPrincipal;
+import com.e2i.wemeet.security.token.TokenInjector;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
