+++ src/main/java/com/e2i/wemeet/security/manager/CreditAuthorize.java
@@ -1,4 +1,4 @@
-package com.e2i.wemeet.config.security.manager;
+package com.e2i.wemeet.security.manager;
 
 import com.e2i.wemeet.domain.member.Role;
 import java.lang.annotation.ElementType;
@@ -7,9 +7,9 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /*
-* Credit 검증 어노테이션
-* */
-@Target({ ElementType.METHOD, ElementType.TYPE })
+ * Credit 검증 어노테이션
+ * */
+@Target({ElementType.METHOD, ElementType.TYPE})
 @Retention(RetentionPolicy.RUNTIME)
 public @interface CreditAuthorize {
