+++ src/main/java/com/e2i/wemeet/security/manager/IsAdmin.java
@@ -1,4 +1,4 @@
-package com.e2i.wemeet.config.security.manager;
+package com.e2i.wemeet.security.manager;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -6,7 +6,7 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 import org.springframework.security.access.prepost.PreAuthorize;
 
-@Target({ ElementType.METHOD, ElementType.TYPE })
+@Target({ElementType.METHOD, ElementType.TYPE})
 @Retention(RetentionPolicy.RUNTIME)
 @PreAuthorize("hasRole('ADMIN')")
 public @interface IsAdmin {
