+++ src/main/java/com/e2i/wemeet/security/token/handler/AccessTokenHandler.java
@@ -1,10 +1,10 @@
-package com.e2i.wemeet.config.security.token.handler;
+package com.e2i.wemeet.security.token.handler;
 
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.algorithms.Algorithm;
 import com.auth0.jwt.interfaces.DecodedJWT;
-import com.e2i.wemeet.config.security.token.JwtEnv;
-import com.e2i.wemeet.config.security.token.Payload;
+import com.e2i.wemeet.security.token.JwtEnv;
+import com.e2i.wemeet.security.token.Payload;
 import java.util.Date;
 import org.springframework.stereotype.Component;
 import org.springframework.util.StringUtils;
