+++ src/main/java/com/e2i/wemeet/security/token/handler/RefreshTokenHandler.java
@@ -1,15 +1,15 @@
-package com.e2i.wemeet.config.security.token.handler;
+package com.e2i.wemeet.security.token.handler;
 
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.algorithms.Algorithm;
-import com.e2i.wemeet.config.security.token.JwtEnv;
-import com.e2i.wemeet.config.security.token.Payload;
+import com.e2i.wemeet.security.token.JwtEnv;
+import com.e2i.wemeet.security.token.Payload;
 import java.util.Date;
 import org.springframework.stereotype.Component;
 
 /*
-* RefreshToken 생성, 검증 로직 수행
-* */
+ * RefreshToken 생성, 검증 로직 수행
+ * */
 @Component
 public class RefreshTokenHandler extends TokenHandler {
