+++ src/main/java/com/e2i/wemeet/security/token/handler/TokenHandler.java
@@ -1,4 +1,4 @@
-package com.e2i.wemeet.config.security.token.handler;
+package com.e2i.wemeet.security.token.handler;
 
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.JWTVerifier;
@@ -8,21 +8,22 @@ import com.auth0.jwt.exceptions.IncorrectClaimException;
 import com.auth0.jwt.exceptions.JWTDecodeException;
 import com.auth0.jwt.exceptions.TokenExpiredException;
 import com.auth0.jwt.interfaces.DecodedJWT;
-import com.e2i.wemeet.config.security.token.JwtEnv;
-import com.e2i.wemeet.config.security.token.Payload;
 import com.e2i.wemeet.exception.ErrorCode;
 import com.e2i.wemeet.exception.token.JwtClaimIncorrectException;
 import com.e2i.wemeet.exception.token.JwtDecodeException;
 import com.e2i.wemeet.exception.token.JwtExpiredException;
 import com.e2i.wemeet.exception.token.JwtSignatureMismatchException;
+import com.e2i.wemeet.security.token.JwtEnv;
+import com.e2i.wemeet.security.token.Payload;
 import java.util.Date;
 import org.springframework.beans.factory.annotation.Value;
 
 /*
-* JWT 관련 역할 수행
-* 토큰 생성 로직은 상속받는 각 클래스에서 구현
-* */
+ * JWT 관련 역할 수행
+ * 토큰 생성 로직은 상속받는 각 클래스에서 구현
+ * */
 public abstract class TokenHandler {
+
     protected static final String ACCESS_PREFIX = "Bearer ";
     protected static final String ISSUER = "WE:MEET";
 
@@ -35,9 +36,9 @@ public abstract class TokenHandler {
     }
 
     /*
-    * JWT 를 검증 및 디코딩
-    * 토큰 검증 단계에서 예외 발생시 CustomException 으로 throw
-    * */
+     * JWT 를 검증 및 디코딩
+     * 토큰 검증 단계에서 예외 발생시 CustomException 으로 throw
+     * */
     public DecodedJWT decodeToken(String token) {
         JWTVerifier verifier = JWT.require(Algorithm.HMAC512(secretKey)).build();
