+++ src/test/java/com/e2i/wemeet/config/security/token/AccessTokenHandlerTest.java
@@ -3,8 +3,9 @@ package com.e2i.wemeet.config.security.token;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import com.e2i.wemeet.config.security.token.handler.AccessTokenHandler;
 import com.e2i.wemeet.domain.member.Role;
+import com.e2i.wemeet.security.token.Payload;
+import com.e2i.wemeet.security.token.handler.AccessTokenHandler;
 import com.e2i.wemeet.support.config.AbstractIntegrationTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
