+++ src/test/java/com/e2i/wemeet/config/security/token/JwtEnvTest.java
@@ -4,6 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import com.e2i.wemeet.domain.member.Role;
+import com.e2i.wemeet.security.token.JwtEnv;
+import com.e2i.wemeet.security.token.Payload;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
